requirements.txt
langgraph==0.6.8
langchain-core==0.3.78
langchain-google-genai==2.1.12
langchain-openai==0.3.35
tqdm==4.67.1
IPython==9.9.0
nltk==3.9.2
sentence-transformers==5.1.1
faiss-cpu==1.12.0
torch==2.8.0
rdflib==7.2.1
spacy==3.8.7
ftfy==6.3.1
pandas==2.3.3
pyyaml==6.0.3
python-dotenv==1.1.1
openpyxl==3.1.5
numpy==2.2.6
pymongo==4.15.2
ijson==3.4.0
dirtyjson==1.0.8
ipykernel==7.1.0
matplotlib==3.10.8